refactor(babel-plugin-espower): move espath generation logic to BabelAssertionVisitor
