+++ lib/babel-assertion-visitor.js
@@ -61,6 +61,17 @@ BabelAssertionVisitor.prototype.enter = function (nodePath) {
     var enclosingFunc = this.findEnclosingFunction(nodePath);
     this.withinGenerator = enclosingFunc && enclosingFunc.generator;
     this.withinAsync = enclosingFunc && enclosingFunc.async;
+    // store original espath for each node
+    var types = this.babel.types;
+    estraverse.traverse(currentNode, {
+        keys: types.VISITOR_KEYS,
+        enter: function (node) {
+            if (this.path()) {
+                var espath = this.path().join('/');
+                define(node, { _espowerEspath: espath });
+            }
+        }
+    });
 };
 
 BabelAssertionVisitor.prototype.enterArgument = function (nodePath) {
@@ -83,7 +94,14 @@ BabelAssertionVisitor.prototype.enterArgument = function (nodePath) {
 };
 
 BabelAssertionVisitor.prototype.leave = function (nodePath) {
-    // nothing to do now
+    var currentNode = nodePath.node;
+    var types = this.babel.types;
+    estraverse.traverse(currentNode, {
+        keys: types.VISITOR_KEYS,
+        enter: function (node) {
+            delete node._espowerEspath;
+        }
+    });
 };
 
 BabelAssertionVisitor.prototype.leaveArgument = function (resultTree) {
