+++ lib/babel-espower-visitor.js
@@ -3,9 +3,7 @@
 var defaultOptions = require('./default-options');
 var CallMatcher = require('call-matcher');
 var babylon = require('babylon');
-var estraverse = require('estraverse');
 var extend = require('xtend');
-var define = require('define-properties');
 var find = require('array-find');
 var BabelAssertionVisitor = require('./babel-assertion-visitor');
 
@@ -45,16 +43,6 @@ BabelEspowerVisitor.prototype.enter = function (nodePath) {
             }, this.options);
             assertionVisitor = new BabelAssertionVisitor(this.babel, matcher, espowerOptions);
             assertionVisitor.enter(nodePath);
-            // mega hack: store original espath for each node
-            estraverse.traverse(currentNode, {
-                keys: types.VISITOR_KEYS,
-                enter: function (node) {
-                    if (this.path()) {
-                        var espath = this.path().join('/');
-                        define(node, { _espowerEspath: espath });
-                    }
-                }
-            });
             file.set('espowerAssertionVisitor', assertionVisitor);
         }
     }
@@ -73,12 +61,6 @@ BabelEspowerVisitor.prototype.exit = function (nodePath) {
         // leaving assertion
         assertionVisitor.leave(nodePath);
         file.delete('espowerAssertionVisitor');
-        estraverse.traverse(currentNode, {
-            keys: types.VISITOR_KEYS,
-            enter: function (node) {
-                delete node._espowerEspath;
-            }
-        });
         return;
     }
     if (!assertionVisitor.isCapturingArgument()) {
